==> Step0_Calculate_Diameter.ipynb <==
import glob
import ntpath
import numpy as np
import pandas as pd


src_dir = "D:\jupyter-notebook\LiverCancer\Image_Nuclei\labels"
csv_target_path = "diameter.csv"

out_csv = []

for i in range(296):
    
    for file_path in glob.glob(src_dir+"/*.csv"):
  
        if int(ntpath.basename(file_path).split(".")[0]) == i:
        
            avg_diameter = np.mean(pd.read_csv(file_path)["diameter"])
            
            if i>=0 and i<=77:
                temp = "ImageN." + str(i+1).zfill(2) + ".jpg"
            elif i>=78 and i<=123:
                temp = "Image1-" + str(i-77).zfill(2) + ".jpg"
            elif i>=124 and i<=176:
                temp = "Image2-" + str(i-123).zfill(2) + ".jpg"
            elif i>=177 and i<=243:
                temp = "Image3-" + str(i-176).zfill(2) + ".jpg"
            elif i<=295:
                temp = "Image4-" + str(i-243).zfill(2) + ".jpg"
            
            csv_line = [temp, avg_diameter]
            print(csv_line)
            out_csv.append(csv_line)
            
df = pd.DataFrame(out_csv, columns=["name", "diameter"])
df.to_csv(csv_target_path, index=False)
