==> Step0_Preprocess_Data_Describe.ipynb <==
import os
import glob
import random
import numpy as np
import pandas as pd

from sklearn.model_selection import KFold

print("Get file_paths.")
        
src_dir = "D:/jupyter-notebook/LiverCancer/Image_Resized/"
file_paths = []

for fold in os.listdir(src_dir):
    file_paths = file_paths + [file_path for file_path in glob.glob(src_dir+fold+"/*.jpg")]

print("Full Count: ",len(file_paths))
random.shuffle(file_paths)
print("Full Count: ",len(file_paths))
file_paths

####################################################################################################
#分割训练数据和测试数据
count = 1
kf=KFold(n_splits=10)
file_paths = np.array(file_paths)

for train_index, holdout_index in kf.split(file_paths):
    
    print("##############################")
    print(count)
    
    train_file_paths = file_paths[train_index]
    holdout_file_paths = file_paths[holdout_index]
    
    columns = ["file_path"]
    df_train = pd.DataFrame(train_file_paths, columns=columns)
    df_holdout = pd.DataFrame(holdout_file_paths, columns=columns)
    
    fold_path = "Data_Description/"+ str(count) + "/"
    
    if not os.path.exists(fold_path):
        os.mkdir(fold_path)
        
    df_train.to_csv(fold_path + "train.csv" , index=False)
    df_holdout.to_csv(fold_path + "holdout.csv" , index=False)
    
    count+=1
    
    print(len(df_train))
    print(len(df_holdout))
