==> Step0_Preprocess_Resize.ipynb <==
import os
import glob
import cv2
import ntpath

print("Get file_paths.")
        
src_dir = "D:/jupyter-notebook/LiverCancer/Image_Mask/"
file_paths = []

for fold in os.listdir(src_dir):
    file_paths = file_paths + [file_path for file_path in glob.glob(src_dir+fold+"/*.jpg")]

dst_path = "D:/jupyter-notebook/LiverCancer/Mask_Resized_512/"

for index, file_path in enumerate(file_paths):
    
    print(index)
    
    sample_image = cv2.imread(file_path)
    sample_image = cv2.resize(sample_image, (512,512)) 
    
    new_file_path = dst_path + file_path.split("/")[-1]
    name_len = len(ntpath.basename(new_file_path))
    fold_path = new_file_path[0:-name_len]

    if not os.path.exists(fold_path):
        os.mkdir(fold_path)
    
    cv2.imwrite(new_file_path, sample_image)
