==> Step1_Preprocess_Generate_Nuclei_Original.ipynb <==
import os
import cv2
import glob
import time
import ntpath
import numpy as np
import pandas as pd

print("Get file_paths.")
        
src_dir = "D:/jupyter-notebook/LiverCancer/Image_Original/"
file_paths = []

for fold in os.listdir(src_dir):
    file_paths = file_paths + [file_path for file_path in glob.glob(src_dir+fold+"/*.jpg")]

dst_path = "D:/jupyter-notebook/LiverCancer/Image_Nuclei/images2/"

for index, file_path in enumerate(file_paths):
    
    time_start=time.time()
    print(file_path)
    
    if "grade0" in file_path:
        sample_label = 0
    elif "grade1" in file_path:
        sample_label = 1
    elif "grade2" in file_path:
        sample_label = 2
    elif "grade3" in file_path:
        sample_label = 3
    elif "grade4" in file_path:
        sample_label = 4

    #####################################################################################
    sample_image = cv2.imread(file_path)
    df_nuclei = pd.read_csv("D:/jupyter-notebook/LiverCancer/Image_Nuclei/labels/" + str(index) + ".csv")
    
    for index, annotation in df_nuclei.iterrows():
        
        index = annotation["index"]
        coord_x = annotation["coord_x"]
        coord_y = annotation["coord_y"]
        diameter = annotation["diameter"] +20
    
        nucleus = get_nucleus_from_image(sample_image, coord_x, coord_y, 128)
        
        #####################################################################################
        fold_name_1 = "grade" + str(sample_label) + "/"
        fold_name_2 = ntpath.basename(file_path) + "/"
        fold_name = fold_name_1 + fold_name_2
        
        if not os.path.exists(dst_path + fold_name):
            os.mkdir(dst_path + fold_name)
            
        #####################################################################################
        number = ntpath.basename(file_path)[-6:-4]
        file_name = "grade_" + str(sample_label) + "_" + str(number)  + "_" + str(index) + "_" + str(coord_x) + "_" + str(coord_y) + "_" + str(diameter-20) + ".jpg"
        image_target_path = dst_path + fold_name + file_name
        cv2.imwrite(image_target_path, nucleus)
        
    #####################################################################################
    time_end=time.time()
    print('totally cost',time_end-time_start)


def get_nucleus_from_image(image, center_x, center_y, box_length):
    
    start_x = max(center_x - box_length / 2, 0)
    if start_x + box_length > image.shape[0]:
        start_x = image.shape[0] - box_length
        
    start_y = max(center_y - box_length / 2, 0)
    if start_y + box_length > image.shape[1]:
        start_y = image.shape[1] - box_length
        
    start_x = int(start_x)
    start_y = int(start_y)

    res = image[start_x:start_x + box_length, start_y:start_y + box_length, :]

    return res

file_path = "D:/jupyter-notebook/LiverCancer/Image_Original/grade0/ImageN.02.jpg"

sample_image = cv2.imread(file_path)
nucleus = get_nucleus_from_image(sample_image, 9, 3932, 61)

cv2.imshow("point", nucleus) #拼接显示为gray
cv2.waitKey(0)
cv2.destroyAllWindows()

src_dir = "D:/jupyter-notebook/LiverCancer/Image_Nuclei/images/"
file_paths = []

for fold_1 in os.listdir(src_dir):
    
    for fold_2 in os.listdir(src_dir+fold_1+"/"):
     
        file_paths = file_paths + [file_path for file_path in glob.glob(src_dir+fold_1+"/"+fold_2+"/*.jpg")]
    
print(len(file_paths))
