==> Step5_Evaluate_Biopsy_Grade_Test.ipynb <==
import numpy as np
import pandas as pd
from collections import Counter

df_1 = pd.read_csv("D:/jupyter-notebook/LiverCancer/Predict_Result/nuclei_level_result_4/1_result.csv")
df_2 = pd.read_csv("D:/jupyter-notebook/LiverCancer/Predict_Result/nuclei_level_result_4/2_result.csv")
df_3 = pd.read_csv("D:/jupyter-notebook/LiverCancer/Predict_Result/nuclei_level_result_4/3_result.csv")
df_4 = pd.read_csv("D:/jupyter-notebook/LiverCancer/Predict_Result/nuclei_level_result_4/4_result.csv")
df_5 = pd.read_csv("D:/jupyter-notebook/LiverCancer/Predict_Result/nuclei_level_result_4/5_result.csv")
df_6 = pd.read_csv("D:/jupyter-notebook/LiverCancer/Predict_Result/nuclei_level_result_4/6_result.csv")
df_7 = pd.read_csv("D:/jupyter-notebook/LiverCancer/Predict_Result/nuclei_level_result_4/7_result.csv")
df_8 = pd.read_csv("D:/jupyter-notebook/LiverCancer/Predict_Result/nuclei_level_result_4/8_result.csv")
df_9 = pd.read_csv("D:/jupyter-notebook/LiverCancer/Predict_Result/nuclei_level_result_4/9_result.csv")
df_10 = pd.read_csv("D:/jupyter-notebook/LiverCancer/Predict_Result/nuclei_level_result_4/10_result.csv")
df = pd.concat([df_1,df_2,df_3,df_4,df_5,df_6,df_7,df_8,df_9,df_10])

results = []
results_softmax = df.iloc[:,1:6].values.tolist()
for i in range(len(df)):
    
    max_index = results_softmax[i].index(max(results_softmax[i]))
    results.append(max_index)
    
file_paths = df.iloc[:,0].values.tolist()

temp1 = np.array(file_paths)[:].reshape(-1,1)
temp2 = np.array(results)[:].reshape(-1,1)

nuclei = np.hstack((temp1, temp2)).tolist()

pec_count = 0

for i in range(10):
    
    current_iteration = i+1

    src_dir = "D:/jupyter-notebook/LiverCancer/Data_Description/" + str(current_iteration) +"/"

    train_biopsy_samples = pd.read_csv(src_dir + "train.csv")["file_path"].tolist()
    holdout_biopsy_samples = pd.read_csv(src_dir + "holdout.csv")["file_path"].tolist()
    print("Train Biopsy Count: ", len(train_biopsy_samples), ", Holdout Biopsy Count: ", len(holdout_biopsy_samples))
    
    ##############################################################################################
    rater_a = []
    rater_b = []
    pec = []
    
    for index_count, holdout_biopsy_sample in enumerate(holdout_biopsy_samples):

        biopsy_name = holdout_biopsy_sample.split(".jpg")[0][-9:]+".jpg"

        if "grade0" in holdout_biopsy_sample:
            tru_label = 0
        elif "grade1" in holdout_biopsy_sample:
            tru_label = 1
        elif "grade2" in holdout_biopsy_sample:
            tru_label = 2
        elif "grade3" in holdout_biopsy_sample:
            tru_label = 3
        elif "grade4" in holdout_biopsy_sample:
            tru_label = 4

        rater_a +=[int(tru_label)]

        nuclei_label = []
        for index, nucleus in enumerate(nuclei):

            if biopsy_name in nucleus[0]:

                nuclei_label += [int(nucleus[1])]

        counter = Counter(nuclei_label)
        print(str(index_count) + ":  " + str(counter))
        pre_label = counter.most_common()[0][0]
        rater_b += [int(pre_label)]
        
        pec += [counter.most_common()[0][1]/len(nuclei_label)]
    
    temp = [a for a in pec if a >0.7]
    pec_count += ( (index_count+1)-len(temp) )

    ##############################################################################################
    print(src_dir)
    print(rater_a)
    print(rater_b)
    
    temp, conf = quadratic_weighted_kappa(rater_a,rater_b)

print(pec_count)

def quadratic_weighted_kappa(rater_a, rater_b, min_rating=None, max_rating=None): 
    
    rater_a = np.array(rater_a, dtype=int) 
    rater_b = np.array(rater_b, dtype=int) 
    
    assert(len(rater_a) == len(rater_b)) 
    
    if min_rating is None:
        min_rating = min(min(rater_a), min(rater_b))
    if max_rating is None:
        max_rating = max(max(rater_a), max(rater_b))
        
    conf_mat = confusion_matrix(rater_a, rater_b, min_rating, max_rating)
    
    print(conf_mat)
    
    num_ratings = len(conf_mat) 
    num_scored_items = float(len(rater_a)) 
    
    hist_rater_a = histogram(rater_a, min_rating, max_rating) 
    hist_rater_b = histogram(rater_b, min_rating, max_rating) 
    
    numerator = 0.0 
    denominator = 0.0 
    
    for i in range(num_ratings): 
        for j in range(num_ratings): 
            
            expected_count = (hist_rater_a[i] * hist_rater_b[j] / num_scored_items) 
            d = pow(i - j, 2.0) / pow(num_ratings - 1, 2.0) 
            numerator += d * conf_mat[i][j] / num_scored_items 
            denominator += d * expected_count / num_scored_items 
            
    return 1.0 - numerator / denominator, conf_mat 

def confusion_matrix(rater_a, rater_b, min_rating=None, max_rating=None):

    assert(len(rater_a) == len(rater_b))

    num_ratings = int(max_rating - min_rating + 1)

    conf_mat = [[0 for i in range(num_ratings)] for j in range(num_ratings)]

    for a, b in zip(rater_a, rater_b):

        conf_mat[a - min_rating][b - min_rating] += 1

    return conf_mat

def histogram(ratings, min_rating=None, max_rating=None):

    num_ratings = int(max_rating - min_rating + 1)

    hist_ratings = [0 for x in range(num_ratings)]

    for r in ratings:

        hist_ratings[r - min_rating] += 1
    
    return hist_ratings
